# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Mini Moodle API server.",
        "title": "Mini Moodle API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/courses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of courses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "List courses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Take",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Courses list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/course.Course"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new course (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Create course",
                "parameters": [
                    {
                        "description": "Create Course Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created course",
                        "schema": {
                            "$ref": "#/definitions/course.Course"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/courses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns course details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Get course by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details",
                        "schema": {
                            "$ref": "#/definitions/course.Course"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update course details (Teacher/Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Update course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated course",
                        "schema": {
                            "$ref": "#/definitions/course.Course"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a course by ID (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Delete course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/student": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets an overview of courses, assignments, and grades for the student",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get student dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dashboard.StudentDashboard"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/student/courses/{courseID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets detailed statistics like attendance and grades for a specific course",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get student course stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dashboard.CourseProgress"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/teacher": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets an overview of courses, students, and grading activities for the teacher",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get teacher dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dashboard.TeacherDashboard"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "course.Course": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "maxPoints": {
                    "type": "integer"
                },
                "teacherID": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "dashboard.CourseProgress": {
            "type": "object",
            "properties": {
                "completed_assignments": {
                    "type": "integer"
                },
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "current_grade": {
                    "type": "number"
                },
                "progress_percent": {
                    "type": "number"
                },
                "total_assignments": {
                    "type": "integer"
                }
            }
        },
        "dashboard.CourseStats": {
            "type": "object",
            "properties": {
                "attendance_rate": {
                    "type": "number"
                },
                "average_grade": {
                    "type": "number"
                },
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "student_count": {
                    "type": "integer"
                },
                "submission_rate": {
                    "type": "number"
                }
            }
        },
        "dashboard.GradeTrend": {
            "type": "object",
            "properties": {
                "average_gpa": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "dashboard.RecentGrade": {
            "type": "object",
            "properties": {
                "course_name": {
                    "type": "string"
                },
                "graded_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_points": {
                    "type": "number"
                },
                "percentage": {
                    "type": "number"
                },
                "score": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "assignment, quiz",
                    "type": "string"
                }
            }
        },
        "dashboard.RecentSubmission": {
            "type": "object",
            "properties": {
                "assignment_title": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "graded": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "student_name": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                }
            }
        },
        "dashboard.StudentDashboard": {
            "type": "object",
            "properties": {
                "active_courses": {
                    "type": "integer"
                },
                "attendance_rate": {
                    "type": "number"
                },
                "average_quiz_score": {
                    "type": "number"
                },
                "completed_assignments": {
                    "type": "integer"
                },
                "course_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.CourseProgress"
                    }
                },
                "gpa": {
                    "type": "number"
                },
                "grade_trends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.GradeTrend"
                    }
                },
                "pending_assignments": {
                    "type": "integer"
                },
                "recent_grades": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.RecentGrade"
                    }
                },
                "student_id": {
                    "type": "string"
                },
                "student_name": {
                    "type": "string"
                },
                "total_courses": {
                    "type": "integer"
                },
                "total_quizzes": {
                    "type": "integer"
                },
                "upcoming_deadlines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.UpcomingDeadline"
                    }
                }
            }
        },
        "dashboard.TeacherDashboard": {
            "type": "object",
            "properties": {
                "course_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.CourseStats"
                    }
                },
                "pending_appeals": {
                    "type": "integer"
                },
                "pending_submissions": {
                    "type": "integer"
                },
                "recent_submissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.RecentSubmission"
                    }
                },
                "teacher_id": {
                    "type": "string"
                },
                "teacher_name": {
                    "type": "string"
                },
                "total_courses": {
                    "type": "integer"
                },
                "total_students": {
                    "type": "integer"
                }
            }
        },
        "dashboard.UpcomingDeadline": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "days_left": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "assignment, quiz",
                    "type": "string"
                }
            }
        },
...
        "http.CreateCourseRequest": {
            "type": "object",
            "required": [
                "code",
                "max_points",
                "teacher_id",
                "title"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "max_points": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "http.UpdateCourseRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "max_points": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}